Move ExplicitRungeKuttaSO3Mixin to integrators.common
